Add nft-asset activity page. Update Zcash.
