+++ UnstoppableWallet/UnstoppableWallet/Modules/NftAsset/NftAssetModule.swift
@@ -1,15 +1,33 @@
 import UIKit
+import LanguageKit
+import MarketKit
 
 struct NftAssetModule {
 
     static func viewController(collectionUid: String, contractAddress: String, tokenId: String, imageRatio: CGFloat) -> UIViewController {
-        let coinPriceService = WalletCoinPriceService(currencyKit: App.shared.currencyKit, marketKit: App.shared.marketKit)
-        let service = NftAssetService(collectionUid: collectionUid, contractAddress: contractAddress, tokenId: tokenId, marketKit: App.shared.marketKit, coinPriceService: coinPriceService)
+        let overviewController = NftAssetOverviewModule.viewController(collectionUid: collectionUid, contractAddress: contractAddress, tokenId: tokenId, imageRatio: imageRatio)
+        let activityController = NftActivityModule.viewController(eventListType: .asset(contractAddress: contractAddress, tokenId: tokenId))
 
-        coinPriceService.delegate = service
+        return NftAssetViewController(
+                overviewController: overviewController,
+                activityController: activityController
+        )
+    }
 
-        let viewModel = NftAssetViewModel(service: service)
-        return NftAssetViewController(viewModel: viewModel, urlManager: UrlManager(inApp: true), imageRatio: imageRatio)
+}
+
+extension NftAssetModule {
+
+    enum Tab: Int, CaseIterable {
+        case overview
+        case activity
+
+        var title: String {
+            switch self {
+            case .overview: return "nft_asset.tab.overview".localized
+            case .activity: return "nft_asset.tab.activity".localized
+            }
+        }
     }
 
 }
