+++ UnstoppableWallet/UnstoppableWallet/Modules/NftAsset/TraitCell.swift
@@ -55,7 +55,7 @@ class TraitCell: UICollectionViewCell {
         fatalError("not implemented")
     }
 
-    func bind(viewItem: NftAssetViewModel.TraitViewItem) {
+    func bind(viewItem: NftAssetOverviewViewModel.TraitViewItem) {
         valueLabel.text = viewItem.value
         typeLabel.text = viewItem.type
         percentBadge.text = viewItem.percent
@@ -63,7 +63,7 @@ class TraitCell: UICollectionViewCell {
         percentBadge.isHidden = viewItem.percent == nil
     }
 
-    static func size(for viewItem: NftAssetViewModel.TraitViewItem, containerWidth: CGFloat) -> CGSize {
+    static func size(for viewItem: NftAssetOverviewViewModel.TraitViewItem, containerWidth: CGFloat) -> CGSize {
         let availableWidth = containerWidth - 2 * horizontalPadding
 
         let badgeWidth = (viewItem.percent.map { BadgeView.width(for: $0, change: nil, style: .small) }) ?? 0
