+++ UnstoppableWallet/UnstoppableWallet/de.lproj/Localizable.strings
@@ -1161,17 +1161,17 @@ Wir haben Unstoppable für uns selbst entwickelt, da die verfügbaren Lösungen 
 "nft_collection.overview.floor_price" = "Mindestpreis";
 "nft_collection.overview.all_time_average" = "Ganzzeitiger Durchschnitt";
 
-"nft_collection.activity.event_types" = "Ereignistypen";
-"nft_collection.activity.event_type.all" = "Alle Ereignisse";
-"nft_collection.activity.event_type.list" = "Liste";
-"nft_collection.activity.event_type.sale" = "Kauf";
-"nft_collection.activity.event_type.offer" = "Angebot eingegeben";
-"nft_collection.activity.event_type.bid" = "Gebot eingegeben";
-"nft_collection.activity.event_type.bid_cancel" = "Gebot zurückgezogen";
-"nft_collection.activity.event_type.transfer" = "Überweisung";
-"nft_collection.activity.event_type.approve" = "Bestätigen";
-"nft_collection.activity.event_type.custom" = "Eigene";
-"nft_collection.activity.event_type.payout" = "Auszahlung";
-"nft_collection.activity.event_type.cancel" = "Abbrechen";
-"nft_collection.activity.event_type.bulkCancel" = "Massenabbruch";
-"nft_collection.activity.event_type.unknown" = "Unbekannt";
+"nft.activity.event_types" = "Ereignistypen";
+"nft.activity.event_type.all" = "Alle Ereignisse";
+"nft.activity.event_type.list" = "Liste";
+"nft.activity.event_type.sale" = "Kauf";
+"nft.activity.event_type.offer" = "Angebot eingegeben";
+"nft.activity.event_type.bid" = "Gebot eingegeben";
+"nft.activity.event_type.bid_cancel" = "Gebot zurückgezogen";
+"nft.activity.event_type.transfer" = "Überweisung";
+"nft.activity.event_type.approve" = "Bestätigen";
+"nft.activity.event_type.custom" = "Eigene";
+"nft.activity.event_type.payout" = "Auszahlung";
+"nft.activity.event_type.cancel" = "Abbrechen";
+"nft.activity.event_type.bulkCancel" = "Massenabbruch";
+"nft.activity.event_type.unknown" = "Unbekannt";
