+++ UnstoppableWallet/UnstoppableWallet/es.lproj/Localizable.strings
@@ -1162,17 +1162,17 @@ poner en peligro la seguridad de mis fondos.";
 "nft_collection.overview.floor_price" = "Precio del piso";
 "nft_collection.overview.all_time_average" = "Promedio de tiempo";
 
-"nft_collection.activity.event_types" = "Tipos de Evento";
-"nft_collection.activity.event_type.all" = "Todos los eventos";
-"nft_collection.activity.event_type.list" = "Lista";
-"nft_collection.activity.event_type.sale" = "Venta";
-"nft_collection.activity.event_type.offer" = "Oferta introducida";
-"nft_collection.activity.event_type.bid" = "Oferta introducida";
-"nft_collection.activity.event_type.bid_cancel" = "Oferta retirada";
-"nft_collection.activity.event_type.transfer" = "Transferir";
-"nft_collection.activity.event_type.approve" = "Aprobar";
-"nft_collection.activity.event_type.custom" = "Personalizado";
-"nft_collection.activity.event_type.payout" = "Pagos";
-"nft_collection.activity.event_type.cancel" = "Cancelar";
-"nft_collection.activity.event_type.bulkCancel" = "Cancelar en masa";
-"nft_collection.activity.event_type.unknown" = "Desconocido";
+"nft.activity.event_types" = "Tipos de Evento";
+"nft.activity.event_type.all" = "Todos los eventos";
+"nft.activity.event_type.list" = "Lista";
+"nft.activity.event_type.sale" = "Venta";
+"nft.activity.event_type.offer" = "Oferta introducida";
+"nft.activity.event_type.bid" = "Oferta introducida";
+"nft.activity.event_type.bid_cancel" = "Oferta retirada";
+"nft.activity.event_type.transfer" = "Transferir";
+"nft.activity.event_type.approve" = "Aprobar";
+"nft.activity.event_type.custom" = "Personalizado";
+"nft.activity.event_type.payout" = "Pagos";
+"nft.activity.event_type.cancel" = "Cancelar";
+"nft.activity.event_type.bulkCancel" = "Cancelar en masa";
+"nft.activity.event_type.unknown" = "Desconocido";
