+++ UnstoppableWallet/UnstoppableWallet/fr.lproj/Localizable.strings
@@ -1161,17 +1161,17 @@ Nous avons construit Unstoppable à notre intention car les solutions disponible
 "nft_collection.overview.floor_price" = "Prix plancher";
 "nft_collection.overview.all_time_average" = "Moyenne du temps";
 
-"nft_collection.activity.event_types" = "Types d'évènement";
-"nft_collection.activity.event_type.all" = "Tous les événements";
-"nft_collection.activity.event_type.list" = "Liste";
-"nft_collection.activity.event_type.sale" = "Vente";
-"nft_collection.activity.event_type.offer" = "Offre saisie";
-"nft_collection.activity.event_type.bid" = "Enchère saisie";
-"nft_collection.activity.event_type.bid_cancel" = "Enchère retirée";
-"nft_collection.activity.event_type.transfer" = "Transfert";
-"nft_collection.activity.event_type.approve" = "Approuver";
-"nft_collection.activity.event_type.custom" = "Personnalisé";
-"nft_collection.activity.event_type.payout" = "Versement";
-"nft_collection.activity.event_type.cancel" = "Annuler";
-"nft_collection.activity.event_type.bulkCancel" = "Annuler en bloc";
-"nft_collection.activity.event_type.unknown" = "Inconnu";
+"nft.activity.event_types" = "Types d'évènement";
+"nft.activity.event_type.all" = "Tous les événements";
+"nft.activity.event_type.list" = "Liste";
+"nft.activity.event_type.sale" = "Vente";
+"nft.activity.event_type.offer" = "Offre saisie";
+"nft.activity.event_type.bid" = "Enchère saisie";
+"nft.activity.event_type.bid_cancel" = "Enchère retirée";
+"nft.activity.event_type.transfer" = "Transfert";
+"nft.activity.event_type.approve" = "Approuver";
+"nft.activity.event_type.custom" = "Personnalisé";
+"nft.activity.event_type.payout" = "Versement";
+"nft.activity.event_type.cancel" = "Annuler";
+"nft.activity.event_type.bulkCancel" = "Annuler en bloc";
+"nft.activity.event_type.unknown" = "Inconnu";
