+++ UnstoppableWallet/UnstoppableWallet/ko.lproj/Localizable.strings
@@ -1161,17 +1161,17 @@ Unstoppable 지갑은 완전한 오픈소스여서 누구나 앱이 요구하는
 "nft_collection.overview.floor_price" = "바닥 가격";
 "nft_collection.overview.all_time_average" = "모든 시간 평균";
 
-"nft_collection.activity.event_types" = "이벤트 유형";
-"nft_collection.activity.event_type.all" = "모든 이벤트";
-"nft_collection.activity.event_type.list" = "목록";
-"nft_collection.activity.event_type.sale" = "판매";
-"nft_collection.activity.event_type.offer" = "입력된 제안";
-"nft_collection.activity.event_type.bid" = "입력된 입찰가";
-"nft_collection.activity.event_type.bid_cancel" = "입찰 철회";
-"nft_collection.activity.event_type.transfer" = "전송";
-"nft_collection.activity.event_type.approve" = "승인";
-"nft_collection.activity.event_type.custom" = "사용자 지정";
-"nft_collection.activity.event_type.payout" = "지불";
-"nft_collection.activity.event_type.cancel" = "취소";
-"nft_collection.activity.event_type.bulkCancel" = "대량 취소";
-"nft_collection.activity.event_type.unknown" = "알 수 없음";
+"nft.activity.event_types" = "이벤트 유형";
+"nft.activity.event_type.all" = "모든 이벤트";
+"nft.activity.event_type.list" = "목록";
+"nft.activity.event_type.sale" = "판매";
+"nft.activity.event_type.offer" = "입력된 제안";
+"nft.activity.event_type.bid" = "입력된 입찰가";
+"nft.activity.event_type.bid_cancel" = "입찰 철회";
+"nft.activity.event_type.transfer" = "전송";
+"nft.activity.event_type.approve" = "승인";
+"nft.activity.event_type.custom" = "사용자 지정";
+"nft.activity.event_type.payout" = "지불";
+"nft.activity.event_type.cancel" = "취소";
+"nft.activity.event_type.bulkCancel" = "대량 취소";
+"nft.activity.event_type.unknown" = "알 수 없음";
