+++ UnstoppableWallet/UnstoppableWallet/ru.lproj/Localizable.strings
@@ -1163,17 +1163,17 @@
 "nft_collection.overview.floor_price" = "Минимальная цена";
 "nft_collection.overview.all_time_average" = "Средняя оценка за все время";
 
-"nft_collection.activity.event_types" = "Типы событий";
-"nft_collection.activity.event_type.all" = "Все события";
-"nft_collection.activity.event_type.list" = "Список";
-"nft_collection.activity.event_type.sale" = "Продажа";
-"nft_collection.activity.event_type.offer" = "Предложение сделано";
-"nft_collection.activity.event_type.bid" = "Ставка сделана";
-"nft_collection.activity.event_type.bid_cancel" = "Ставка отозвана";
-"nft_collection.activity.event_type.transfer" = "Перевод";
-"nft_collection.activity.event_type.approve" = "Разрешить";
-"nft_collection.activity.event_type.custom" = "Индивидуальная";
-"nft_collection.activity.event_type.payout" = "Выплата";
-"nft_collection.activity.event_type.cancel" = "Отмена";
-"nft_collection.activity.event_type.bulkCancel" = "Массовая отмена";
-"nft_collection.activity.event_type.unknown" = "Неизвестная";
+"nft.activity.event_types" = "Типы событий";
+"nft.activity.event_type.all" = "Все события";
+"nft.activity.event_type.list" = "Список";
+"nft.activity.event_type.sale" = "Продажа";
+"nft.activity.event_type.offer" = "Предложение сделано";
+"nft.activity.event_type.bid" = "Ставка сделана";
+"nft.activity.event_type.bid_cancel" = "Ставка отозвана";
+"nft.activity.event_type.transfer" = "Перевод";
+"nft.activity.event_type.approve" = "Разрешить";
+"nft.activity.event_type.custom" = "Индивидуальная";
+"nft.activity.event_type.payout" = "Выплата";
+"nft.activity.event_type.cancel" = "Отмена";
+"nft.activity.event_type.bulkCancel" = "Массовая отмена";
+"nft.activity.event_type.unknown" = "Неизвестная";
