+++ UnstoppableWallet/UnstoppableWallet/tr.lproj/Localizable.strings
@@ -1164,17 +1164,17 @@ Mevcut çözümler yeterli olmadığından kendimiz için Durdurulamaz'ı geliş
 "nft_collection.overview.floor_price" = "Taban Fiyatı";
 "nft_collection.overview.all_time_average" = "Tüm Zamanların Ortalaması";
 
-"nft_collection.activity.event_types" = "Etkinlik Türleri";
-"nft_collection.activity.event_type.all" = "Tüm Etkinlikler";
-"nft_collection.activity.event_type.list" = "Liste";
-"nft_collection.activity.event_type.sale" = "İndirim";
-"nft_collection.activity.event_type.offer" = "Son Teklif Girildi";
-"nft_collection.activity.event_type.bid" = "Teklif Girildi";
-"nft_collection.activity.event_type.bid_cancel" = "Teklif Geri Çekildi";
-"nft_collection.activity.event_type.transfer" = "Aktar";
-"nft_collection.activity.event_type.approve" = "Onayla";
-"nft_collection.activity.event_type.custom" = "Özel";
-"nft_collection.activity.event_type.payout" = "Ödeme";
-"nft_collection.activity.event_type.cancel" = "İptal";
-"nft_collection.activity.event_type.bulkCancel" = "Toplu İptal";
-"nft_collection.activity.event_type.unknown" = "Bilinmeyen";
+"nft.activity.event_types" = "Etkinlik Türleri";
+"nft.activity.event_type.all" = "Tüm Etkinlikler";
+"nft.activity.event_type.list" = "Liste";
+"nft.activity.event_type.sale" = "İndirim";
+"nft.activity.event_type.offer" = "Son Teklif Girildi";
+"nft.activity.event_type.bid" = "Teklif Girildi";
+"nft.activity.event_type.bid_cancel" = "Teklif Geri Çekildi";
+"nft.activity.event_type.transfer" = "Aktar";
+"nft.activity.event_type.approve" = "Onayla";
+"nft.activity.event_type.custom" = "Özel";
+"nft.activity.event_type.payout" = "Ödeme";
+"nft.activity.event_type.cancel" = "İptal";
+"nft.activity.event_type.bulkCancel" = "Toplu İptal";
+"nft.activity.event_type.unknown" = "Bilinmeyen";
