+++ UnstoppableWallet/UnstoppableWallet/zh.lproj/Localizable.strings
@@ -1163,17 +1163,17 @@ Unstoppable钱包完全开源且任何人都可以确定应用按照其宣称的
 "nft_collection.overview.floor_price" = "底价";
 "nft_collection.overview.all_time_average" = "全部时间平均值";
 
-"nft_collection.activity.event_types" = "事件类型";
-"nft_collection.activity.event_type.all" = "全部事件";
-"nft_collection.activity.event_type.list" = "列表";
-"nft_collection.activity.event_type.sale" = "出售";
-"nft_collection.activity.event_type.offer" = "已输入报价";
-"nft_collection.activity.event_type.bid" = "已输入出价";
-"nft_collection.activity.event_type.bid_cancel" = "已撤回出价";
-"nft_collection.activity.event_type.transfer" = "转帐";
-"nft_collection.activity.event_type.approve" = "批准";
-"nft_collection.activity.event_type.custom" = "自定义";
-"nft_collection.activity.event_type.payout" = "支出";
-"nft_collection.activity.event_type.cancel" = "取消";
-"nft_collection.activity.event_type.bulkCancel" = "批量取消";
-"nft_collection.activity.event_type.unknown" = "未知";
+"nft.activity.event_types" = "事件类型";
+"nft.activity.event_type.all" = "全部事件";
+"nft.activity.event_type.list" = "列表";
+"nft.activity.event_type.sale" = "出售";
+"nft.activity.event_type.offer" = "已输入报价";
+"nft.activity.event_type.bid" = "已输入出价";
+"nft.activity.event_type.bid_cancel" = "已撤回出价";
+"nft.activity.event_type.transfer" = "转帐";
+"nft.activity.event_type.approve" = "批准";
+"nft.activity.event_type.custom" = "自定义";
+"nft.activity.event_type.payout" = "支出";
+"nft.activity.event_type.cancel" = "取消";
+"nft.activity.event_type.bulkCancel" = "批量取消";
+"nft.activity.event_type.unknown" = "未知";
